+++ delta/app/src/main/java/com/example/delta/AccelLoggerService.kt
@@ -11,7 +11,9 @@ import android.hardware.SensorEventListener
 import android.hardware.SensorManager
 import android.os.IBinder
 import android.util.Log
-import androidx.core.app.ActivityCompat.startActivityForResult
+import java.io.File
+import java.io.FileOutputStream
+import java.text.SimpleDateFormat
 import java.util.*
 
 
@@ -23,9 +25,21 @@ class AccelLoggerService: Service(), SensorEventListener {
     private val LAUNCH_END_BUTTON_CODE = 2
     private lateinit var chosenActivity: String
 
+    private lateinit var dataFolderName: String
+    private lateinit var fRaw: FileOutputStream
+    private lateinit var rawFilename: String
+    private lateinit var fSession: FileOutputStream
+    private lateinit var sessionFilename: String
+    private var rawFileIndex: Int = 0
+    private var currentActivity: String = "None"
+//    private var calendar = Calendar.getInstance()
+//    private var startTimeMillis = Calendar.getInstance().timeInMillis
+    private val startTimeReadable = SimpleDateFormat("yyyy-MM-dd_HH_mm_ss", Locale.ENGLISH).format(Date())
+
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
         super.onStartCommand(intent, flags, startId)
         chosenActivity = getString(R.string.NO_ACTIVITY)
+        createFiles()
 
         val intentFilter = IntentFilter(getString(R.string.BROADCAST_CODE))
         activityChangeReceiver = ActivityChangeReceiver()
@@ -79,17 +93,47 @@ class AccelLoggerService: Service(), SensorEventListener {
             .setTicker("Ticker")
             .build()
     }
+    private fun createFiles(){
+        // create folder for this session's files
+        dataFolderName = startTimeReadable
+        File(this.filesDir, dataFolderName).mkdir()
+        createNewRawFile()
+
+        sessionFilename = "Session.$startTimeReadable.csv"    // file to save session information
+        fSession = FileOutputStream(File(this.filesDir, "$dataFolderName/$sessionFilename"))
+        writeToSessionFile("File Start Time: ${Calendar.getInstance().timeInMillis}\n")
+        writeToSessionFile("Event,Start Time,Stop Time\n")
+    }
+    private fun createNewRawFile() {
+        Log.i("0003", "Creating New Raw File")
+        if (rawFileIndex != 0) {
+            fRaw.close()
+        }
+        rawFilename = "$startTimeReadable.$rawFileIndex.csv"       // file to save raw data
+        fRaw = FileOutputStream(File(this.filesDir, "$dataFolderName/$rawFilename"))
+            fRaw.write("File Start Time: ${Calendar.getInstance().timeInMillis}\n".toByteArray())
+        fRaw.write("timestamp,acc_x,acc_y,acc_z,real time,activity\n".toByteArray())
+        rawFileIndex++
+    }
+    private fun writeToSessionFile(str: String) {
+        Log.i("0003", "Writing to Session File")
+        fSession = FileOutputStream(File(this.filesDir, "$dataFolderName/$sessionFilename"), true)
+        fSession.use { f ->
+            f.write(str.toByteArray())
+        }
+    }
     private inner class ActivityChangeReceiver : BroadcastReceiver() {
         override fun onReceive(context: Context?, intent: Intent) {
             if (intent.action == getString(R.string.BROADCAST_CODE)) {
                 val chosenActivity = intent.getStringExtra(getString(R.string.ACTIVITY))
                 if(chosenActivity == getString(R.string.NO_ACTIVITY)){
-                    Log.i("0003", "Ended $chosenActivity}")
-
+                    Log.i("0003", "Ended Activity")
+                    writeToSessionFile("${Calendar.getInstance().timeInMillis}\n")
+                    createNewRawFile()
                 }
                 else{
-                    Log.i("0003", "Started $chosenActivity}")
-                    
+                    Log.i("0003", "Started $chosenActivity")
+                    writeToSessionFile("$chosenActivity,${Calendar.getInstance().timeInMillis},")
                 }
             }
         }
